Refactoring log_model to deployable_entity.
